refactor: Moved Assignment components to own files
