+++ src/pages/assignments.tsx
@@ -1,58 +1,15 @@
 // Modules
-import { format, formatDuration, intervalToDuration, isPast } from "date-fns";
 import { NextPage } from "next";
 import { useState } from "react";
-import ReactMarkdown from "react-markdown";
 
 // Components
-import Status from "@components/assignments/status";
-import Headline from "@components/global/headline";
-import MaterialIcon from "@components/global/icon/material";
-import Title from "@components/global/title";
-import FilterChip from "@components/input/chip/filter";
-import ChoiceChip from "@components/input/chip/choice";
+import AssignmentDetails from "@components/assignments/details"
+import AssignmentItem from "@components/assignments/item";
 import Search from "@components/input/search";
 import TopAppBar from "@components/layout/top-app-bar";
 
 // Types
 import { Assignment } from "@utils/types/assignment";
-
-const AssignmentItem = ({
-  assignment,
-  active,
-  setActiveID,
-}: {
-  assignment: Assignment;
-  active: boolean;
-  setActiveID: Function;
-}): JSX.Element => {
-  return (
-    <li className={`list-page-list-item ${active ? "active" : "not-active"}`}>
-      <button
-        className="flex flex-col gap-2 sm:flex-row sm:justify-between sm:items-end p-4 sm:h-20"
-        onClick={() => {
-          setActiveID(assignment.id);
-        }}
-      >
-        <div className="flex flex-col">
-          <p className="font-bold max-lines-1">{assignment.name}</p>
-          <p className="max-lines-1">
-            {assignment.subject}
-            {" • "}
-            {format(assignment.due, "dd/MM/yyyy")}
-          </p>
-        </div>
-        <div className="flex flex-row w-full justify-end sm:w-fit">
-          <Status
-            status={assignment.status}
-            urgent={assignment.urgent}
-            pastDue={isPast(assignment.due)}
-          />
-        </div>
-      </button>
-    </li>
-  );
-};
 
 const ActiveAssignmentDisplay = ({
   assignment,
@@ -61,133 +18,7 @@ const ActiveAssignmentDisplay = ({
 }) => {
   return (
     <div className="card card-elevated">
-      <Title title={<h2>{assignment.name}</h2>} subhead={assignment.subject} />
-      <div className="flex flex-col sm:grid sm:grid-cols-2">
-        <Title
-          icon={
-            <MaterialIcon
-              icon="calendar_today"
-              className="text-light-primary dark:text-dark-primary"
-            />
-          }
-          title={<h3>Due Date</h3>}
-          subhead={format(assignment.due, "dd/MM/yyyy HH:mm")}
-        />
-        <Title
-          icon={
-            <MaterialIcon
-              icon={
-                isPast(assignment.due)
-                  ? assignment.status == "done"
-                    ? "hourglass_bottom"
-                    : "assignment_late"
-                  : "hourglass_top"
-              }
-              className="text-light-primary dark:text-dark-primary"
-            />
-          }
-          title={
-            <h3>
-              {isPast(assignment.due)
-                ? assignment.status == "done"
-                  ? "Was Due"
-                  : "Past Due"
-                : "Time Left"}
-            </h3>
-          }
-          subhead={`${formatDuration(
-            intervalToDuration({ start: assignment.due, end: new Date() }),
-            { format: ["years", "months", "days", "hours"] }
-          )} ${isPast(assignment.due) ? "ago" : "left"}`}
-        />
-      </div>
-      <Headline
-        icon={
-          <MaterialIcon
-            icon="assignment"
-            className="self-center text-light-primary dark:text-dark-primary"
-          />
-        }
-        title={<h3 className="w-full text-xl font-medium">Status</h3>}
-        subhead={
-          <div className="flex flex-row gap-2">
-            <ChoiceChip
-              choices={[
-                {
-                  id: "not-started",
-                  name: "Not started",
-                  style: {
-                    active: "container-tertiary",
-                    inactive:
-                      "transition-colors \
-                      hover:bg-light-tertiary-0.08-tlc hover:dark:bg-dark-tertiary-0.08-tlc transition-none \
-                      focus:bg-light-tertiary-0.12-tlc focus:dark:bg-dark-tertiary-0.12-tlc ring-0",
-                  },
-                },
-                {
-                  id: "started",
-                  name: "Started",
-                  style: {
-                    active: "container-secondary",
-                    inactive:
-                      "transition-colors \
-                      hover:bg-light-secondary-0.08-tlc hover:dark:bg-dark-secondary-0.08-tlc transition-none \
-                      focus:bg-light-secondary-0.12-tlc focus:dark:bg-dark-secondary-0.12-tlc ring-0",
-                  },
-                },
-                {
-                  id: "done",
-                  name: "Done",
-                  style: {
-                    active: "container-primary",
-                    inactive:
-                      "transition-colors \
-                      hover:bg-light-primary-0.08-tlc hover:dark:bg-dark-primary-0.08-tlc transition-none \
-                      focus:bg-light-primary-0.12-tlc focus:dark:bg-dark-primary-0.12-tlc ring-0",
-                  },
-                },
-              ]}
-              onChange={(id: "not-started" | "started" | "done") => {}}
-              value={assignment.status}
-            />
-            <FilterChip
-              name="Urgent"
-              onClick={(active: boolean) => {}}
-              style={{
-                active: "container-error",
-                inactive:
-                  "transition-colors \
-                  hover:bg-light-error-0.08-tlc hover:dark:bg-dark-error-0.08-tlc transition-none \
-                  focus:bg-light-error-0.12-tlc focus:dark:bg-dark-error-0.12-tlc ring-0",
-              }}
-              value={assignment.urgent}
-            />
-            {isPast(assignment.due) && (
-              <div
-                className="chip text-light-on-error dark:text-dark-on-error
-                  bg-light-error dark:bg-dark-error"
-              >
-                Past Due
-              </div>
-            )}
-          </div>
-        }
-        gap={4}
-      />
-      <ReactMarkdown className="p-4">{assignment.desc}</ReactMarkdown>
-      <div className="flex flex-row justify-end items-center gap-2 p-4">
-        <button className="btn btn-outlined">Compose about</button>
-        {assignment.source && (
-          <a
-            href={assignment.source}
-            target="_blank"
-            rel="noreferrer"
-            className="btn btn-filled"
-          >
-            Open in Classroom
-          </a>
-        )}
-      </div>
+      <AssignmentDetails assignment={assignment} />
     </div>
   );
 };
